requirements.txt
adjustText==0.7.3
gensim==4.2.0
matchms==0.17.0
matplotlib==3.5.3
molmass==2021.6.18
ms2deepscore==0.2.3
networkx==2.8.6
numba==0.56.2
numpy==1.23.3
pandas==1.4.4
pyinstaller==5.5
rdkit-pypi
scikit-learn==1.1.3
scikit-opt==0.6.6
scipy==1.9.3
seaborn==0.12.1
spec2vec==0.6.0
tqdm==4.64.1
PyQt5==5.15.7
hnswlib==0.6.2
JPype1==1.4.1
beautifulsoup4==4.11.1
gradio==4.1.2
requests~=2.31.0
sqlalchemy~=2.0.23
uvicorn~=0.24.0.post1
fastapi~=0.104.1
pyyaml~=6.0.1
rdkit-pypi